Refine plugin architecture boundary review fixes
